+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Wolf-hound now has their allegiance hidden when they are lynched.
 * Fixed the grammar on the first Werewolves' turn messages.
 * Moved Wolf-hound's turn to before the Seer's so that the Seer may see his allegiance properly.
+* Restricted specifying `Simple Villager` or `Simple Werewolf` as extra roles.
 
 #### v0.4.6.0